Create BMSReader_8Pack_WiFiAP.ino
CAN bus reader to monitor 8 Kohler batteries. Uses Access Point (AP) for config and is configurable to home wifi for access. 
AP is at SSID: ESP32_CANBus. The IP to access on AP is 192.168.4.1 or use hostname same as SSID: ESP32_CANBus. The config page is at ESP32_CANBus/config